pyproject.toml
[tool.poetry]
name = "qa-guru-4-diplom"
version = "0.1.0"
description = ""
authors = ["peresypkinmi <peresypkin.ma@sibvaleo.com>"]
readme = "README.md"
packages = [{include = "qa_guru_4_diplom"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.2"
allure-pytest = "^2.13.2"
requests = "^2.31.0"
selene = "2.0.0b17"
curlify = "^2.2.1"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
